Move kerning data related code to own file
